+++ exuetangWeb/src/com/yutai/exuetang/model/dao/exuetang/ICoinsDAO.java
@@ -11,7 +11,7 @@ public interface ICoinsDAO {
 	//查找积分
 	public abstract Coins selectCoinsByCoinsID(int coins_id);
 	//根据用户ID查找积分数量
-	public abstract int selectCoinsNumByUserID(int user_id);
+	public abstract double selectCoinsNumByUserID(int user_id);
 	//根据用户ID和积分更新积分
 	public abstract boolean updateCoinsByUserID(int user_id,int coins_num);
 	//查找所有积分
